Moving entry date validation and adding tests
This moves the validateEntryDate to a separate package and adds unit tests for the date.
